fix(packed): keep lifecycle projection peer-free
Separate React rendering from event envelope projection so importing the public API does not load optional React peers.
