+++ .changeset/packed-runtime-boundaries.md
@@ -2,5 +2,6 @@
 'agent-bundle': patch
 ---
 
-Keep the public API importable without the optional runtime peer, and keep the
-packed event journey's request-context probe within native hook output fields.
+Keep the public API importable without optional runtime or React peers, and
+keep the packed event journey's request-context probe within native hook
+output fields.
